+++ src/main/java/com/example/demo/config/AuthConfig.java
@@ -1,13 +1,11 @@
 package com.example.demo.config;
 
-
 import com.nimbusds.jose.jwk.JWKSet;
 import com.nimbusds.jose.jwk.RSAKey;
 import com.nimbusds.jose.jwk.source.JWKSource;
 import com.nimbusds.jose.proc.SecurityContext;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
@@ -22,17 +20,13 @@ import org.springframework.security.oauth2.server.authorization.config.annotatio
 import org.springframework.security.oauth2.server.authorization.settings.AuthorizationServerSettings;
 import org.springframework.security.web.SecurityFilterChain;
 
-
-
 import java.security.Key;
 import java.util.UUID;
 
 @Configuration
 public class AuthConfig {
-
     @Value("${menstrunation.host}")
     private String host;
-
     private static final String SECRET = "my-secret-key-my-secret-key-my-secret-key!";// must be 256-bit for HS25
 
     private static Key getSigningKey() {
@@ -43,20 +37,21 @@ public class AuthConfig {
     @Bean
     public SecurityFilterChain authorizationServerSecurityFilterChain(HttpSecurity http) throws Exception {
         OAuth2AuthorizationServerConfigurer authorizationServerConfigurer =
-                new OAuth2AuthorizationServerConfigurer();
+            new OAuth2AuthorizationServerConfigurer();
 
         http
-                .authorizeHttpRequests(authorize -> authorize.anyRequest().authenticated())
-                .csrf(csrf -> csrf.ignoringRequestMatchers(authorizationServerConfigurer.getEndpointsMatcher()))
-                .exceptionHandling(exceptions -> exceptions
-                        .authenticationEntryPoint(
-                                new org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint("/login")
-                        )
+            .authorizeHttpRequests(authorize -> authorize.anyRequest().authenticated())
+            .csrf(csrf -> csrf.ignoringRequestMatchers(authorizationServerConfigurer.getEndpointsMatcher()))
+            .exceptionHandling(exceptions -> exceptions
+                .authenticationEntryPoint(
+                    new org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint("/login")
                 )
-                .oauth2ResourceServer(resourceServer -> resourceServer.jwt(Customizer.withDefaults()))
-                .with(authorizationServerConfigurer, configurer -> configurer
-                        .oidc(Customizer.withDefaults()) // Enable OpenID Connect 1.0
-                );
+            )
+            .oauth2ResourceServer(resourceServer -> resourceServer.jwt(Customizer.withDefaults()))
+            .with(
+                authorizationServerConfigurer, configurer -> configurer
+                    .oidc(Customizer.withDefaults()) // Enable OpenID Connect 1.0
+            );
 
         return http.build();
     }
@@ -64,15 +59,15 @@ public class AuthConfig {
     @Bean
     public RegisteredClientRepository registeredClientRepository() {
         RegisteredClient registeredClient = RegisteredClient.withId(UUID.randomUUID().toString())
-                .clientId("client")
-                .clientSecret("{noop}secret")
-                .clientAuthenticationMethod(ClientAuthenticationMethod.CLIENT_SECRET_BASIC)
-                .authorizationGrantType(AuthorizationGrantType.AUTHORIZATION_CODE)
-                .authorizationGrantType(AuthorizationGrantType.REFRESH_TOKEN)
-                .redirectUri("http://localhost:8080/login/oauth2/code/client-oidc")
-                .scope(OidcScopes.OPENID)
-                .scope("read")
-                .build();
+            .clientId("client")
+            .clientSecret("{noop}secret")
+            .clientAuthenticationMethod(ClientAuthenticationMethod.CLIENT_SECRET_BASIC)
+            .authorizationGrantType(AuthorizationGrantType.AUTHORIZATION_CODE)
+            .authorizationGrantType(AuthorizationGrantType.REFRESH_TOKEN)
+            .redirectUri("http://localhost:8080/login/oauth2/code/client-oidc")
+            .scope(OidcScopes.OPENID)
+            .scope("read")
+            .build();
 
         return new InMemoryRegisteredClientRepository(registeredClient);
     }
@@ -87,7 +82,7 @@ public class AuthConfig {
     @Bean
     public AuthorizationServerSettings providerSettings() {
         return AuthorizationServerSettings.builder()
-                .issuer(host)
-                .build();
+            .issuer(host)
+            .build();
     }
 }
