--- src/main/resources/application-dev.yml
@@ -1,13 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: org.mariadb.jdbc.Driver
-    url: jdbc:mariadb://localhost:4304/menstrunation
-    username: root
-    password: root
-  jpa:
-    database-platform: org.hibernate.dialect.MariaDBDialect
-    hibernate:
-      ddl-auto: none
-  flyway:
-    enabled: true
-    locations: classpath:db/migration
