+++ src/main/resources/application.properties
@@ -4,6 +4,6 @@ spring.datasource.username=root
 spring.datasource.password=root
 spring.jpa.database-platform=org.hibernate.dialect.MariaDBDialect
 spring.jpa.hibernate.ddl-auto=none
-
+menstrunation.host=http://localhost:8080
 spring.flyway.enabled=true
 spring.flyway.locations=classpath:db/migration
